requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.2
build==1.3.0
CacheControl==0.14.3
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cleo==2.1.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crashtest==0.4.1
cryptography==46.0.1
decouple==0.0.7
distlib==0.4.0
Django==5.2.6
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
dulwich==0.24.1
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.7.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
kombu==5.5.4
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.29
more-itertools==10.8.0
msgpack==1.1.1
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pbs-installer==2025.9.18
pkginfo==1.12.1.2
platformdirs==4.4.0
poetry-core==2.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
PyYAML==6.0.2
RapidFuzz==3.14.1
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
SecretStorage==3.4.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
tenacity==9.1.2
tomlkit==0.13.3
trove-classifiers==2025.9.11.17
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
xxhash==3.5.0
zstandard==0.25.0
